+++ CHANGELOG.md
@@ -7,7 +7,15 @@ availableAt:
 ---
 
 ### unreleased
-**Git History**: <https://github.com/Noelware/violet/compare/26.05.02...master>
+**Git History**: <https://github.com/Noelware/violet/compare/26.05.03...master>
+
+### 26.05.03
+#### Noelware.Violet.Experimental
+- Don't decay return types in `OneOf::Visit` ([`@auguwu`])
+- Add default constructor for `OneOf` if the first type is default constructible ([`@auguwu`])
+- Add **const&&** overload for `OneOf::Visit` ([`@auguwu`])
+
+**Git History**: <https://github.com/Noelware/violet/compare/26.05.02...26.05.03>
 
 ### 26.05.02
 #### Noelware.Violet